Moved all authentication related files to utilities folder
