+++ src/utilities/PrivateRoute.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Route } from 'react-router-dom';
-import { useAuth0 } from '../../auth-wrapper.js';
-import LandingPage from '../../views/LandingPage/LandingPage.js';
+import { useAuth0 } from './auth-wrapper.js';
+import LandingPage from '../views/LandingPage/LandingPage.js';
 
 const PrivateRoute = ({ component: Component, path, ...rest }) => {
   const { isAuthenticated } = useAuth0();
