+++ src/views/LandingPage/LandingPage.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import { useAuth0 } from '../../auth-wrapper';
+import { useAuth0 } from '../../utilities/auth-wrapper';
 import './SASS/LandingPage.scss';
 
 import x from '../../ASSETS/logo.svg';
